+++ src/Krp/Dns/DnsLookupOptions.cs
@@ -1,0 +1,9 @@
+﻿namespace Krp.Dns;
+
+public class DnsLookupOptions
+{
+    /// <summary>
+    /// DNS server.
+    /// </summary>
+    public string Nameserver { get; set; }
+}
