+++ src/Krp/Dns/DnsOptions.cs
@@ -1,0 +1,9 @@
+﻿namespace Krp.Dns;
+
+public enum DnsOptions
+{
+    /// <summary>
+    /// Use Windows HOSTS file DNS routing.
+    /// </summary>
+    WindowsHostsFile,
+}
