+++ src/Krp/Dns/DnsWindowsHostsOptions.cs
@@ -1,0 +1,9 @@
+﻿namespace Krp.Dns;
+
+public class DnsWindowsHostsOptions
+{
+    /// <summary>
+    /// Path to HOSTS file.
+    /// </summary>
+    public string Path { get; set; }
+}
