+++ src/Krp/Endpoints/Models/HttpEndpoint.cs
@@ -1,0 +1,8 @@
+﻿namespace Krp.Endpoints.Models;
+
+public class HttpEndpoint
+{
+    public string Host { get; set; }
+    public int LocalPort { get; set; }
+    public string Path { get; set; }
+}
