+++ src/Krp/Forwarders/TcpForwarder/TcpForwarderOptions.cs
@@ -1,0 +1,9 @@
+using System.Net;
+
+namespace Krp.Forwarders.TcpForwarder;
+
+public class TcpForwarderOptions
+{
+    public IPAddress ListenAddress { get; set; }
+    public int[] ListenPorts { get; set; }
+}
