--- src/Krp/KubernetesForwarder/Dns/DnsLookupOptions.cs
@@ -1,9 +1,0 @@
-﻿namespace Krp.KubernetesForwarder.Dns;
-
-public class DnsLookupOptions
-{
-    /// <summary>
-    /// DNS server.
-    /// </summary>
-    public string Nameserver { get; set; }
-}
