--- src/Krp/KubernetesForwarder/Dns/DnsOptions.cs
@@ -1,9 +1,0 @@
-﻿namespace Krp.KubernetesForwarder.Dns;
-
-public enum DnsOptions
-{
-    /// <summary>
-    /// Use Windows HOSTS file DNS routing.
-    /// </summary>
-    WindowsHostsFile,
-}
