--- src/Krp/KubernetesForwarder/Dns/DnsWindowsHostsOptions.cs
@@ -1,9 +1,0 @@
-﻿namespace Krp.KubernetesForwarder.Dns;
-
-public class DnsWindowsHostsOptions
-{
-    /// <summary>
-    /// Path to HOSTS file.
-    /// </summary>
-    public string Path { get; set; }
-}
