--- src/Krp/KubernetesForwarder/Endpoints/Models/HttpEndpoint.cs
@@ -1,8 +1,0 @@
-﻿namespace Krp.KubernetesForwarder.Endpoints.Models;
-
-public class HttpEndpoint
-{
-    public string Host { get; set; }
-    public int LocalPort { get; set; }
-    public string Path { get; set; }
-}
