--- src/Krp/KubernetesForwarder/Forwarders/HttpForwarder/Startup.cs
@@ -1,25 +1,0 @@
-using Krp.KubernetesForwarder.Forwarders.HttpForwarder;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.DependencyInjection;
-
-public class Startup
-{
-    public void ConfigureServices(IServiceCollection services)
-    {
-        services.AddHttpForwarder();
-        services.AddSingleton<HttpForwarderHandler>();
-    }
-
-    public void Configure(IApplicationBuilder app)
-    {
-        app.UseRouting();
-        app.UseEndpoints(endpoints =>
-        {
-            endpoints.Map("/{**catch-all}", async (HttpForwarderHandler handler, HttpContext httpContext) =>
-            {
-                await handler.HandleRequest(httpContext);
-            });
-        });
-    }
-}
