--- src/Krp/KubernetesForwarder/Forwarders/TcpForwarder/TcpForwarderOptions.cs
@@ -1,9 +1,0 @@
-using System.Net;
-
-namespace Krp.KubernetesForwarder.Forwarders.TcpForwarder;
-
-public class TcpForwarderOptions
-{
-    public IPAddress ListenAddress { get; set; }
-    public int[] ListenPorts { get; set; }
-}
